test(logging): consolidate the logging tests into one file
Merge test_logging_infra / _phase1_warnings / _phase2_logging / _phase3_cli
into a single tests/test_logging.py, organized by section (infrastructure,
warnings, logging, CLI). Same 29 tests, no behavior change.

The web app's warnings-capture test stays in web/tests/test_logging.py — it
belongs to the separate web test suite (imports the NiceGUI 'main' module
via that suite's conftest/sys.path).
